CollpasedStack use native Renderer
CollapsedStackRenderer implements IIngredientRenderer<>
