+++ src/main/java/mezz/jei/plugins/jei/JEIInternalPlugin.java
@@ -11,7 +11,7 @@ import mezz.jei.bookmarks.BookmarkItem;
 import mezz.jei.bookmarks.BookmarkItemRender;
 import mezz.jei.ingredients.CollapsedStack;
 import mezz.jei.ingredients.CollapsedStackIngredientHelper;
-import mezz.jei.ingredients.CollapsedStackIngredientRenderer;
+import mezz.jei.render.CollapsedStackRenderer;
 import net.minecraftforge.fml.common.registry.ForgeRegistries;
 import net.minecraftforge.fluids.Fluid;
 import net.minecraftforge.fluids.FluidRegistry;
@@ -66,8 +66,7 @@ public class JEIInternalPlugin implements IModPlugin {
 
 		// Register CollapsedStack as ingredient type — addons that introspect grid items require a registered type
 		CollapsedStackIngredientHelper csHelper = new CollapsedStackIngredientHelper();
-		CollapsedStackIngredientRenderer csRenderer = new CollapsedStackIngredientRenderer();
-		ingredientRegistration.register(CollapsedStack.TYPE, Collections.emptyList(), csHelper, csRenderer);
+		ingredientRegistration.register(CollapsedStack.TYPE, Collections.emptyList(), csHelper, CollapsedStackRenderer.INSTANCE);
 	}
 
 	@Override
